Вынесла константы для тестов в отдельный класс, отрефакторила тест на addAds()

+++ src/main/java/ru/skypro/homework/dto/AdsDto.java
@@ -8,7 +8,7 @@ import javax.validation.constraints.NotBlank;
 @Data
 public class AdsDto {
 
-    private int author;
+    private long author;
 
     private String image;
